Create a new xmode "Docs"
This is a compact mode with a minimal header and no traceback,
mosttly to be used in Documentation, where the context of the traceback
is rarely useful

```
$ ipython --InteractiveShell.xmode=Docs

In [1]: import pywt
   ...:
   ...: a = pywt.idwt([1,2,3,4,5], [1,2,3,4], 'db2' ,'symmetric')
   ...:
   ...: print(a)
Traceback (most recent call last): Cell In[1], line 3

->  a = pywt.idwt([1,2,3,4,5], [1,2,3,4], 'db2' ,'symmetric')

    [... 3 skipped frames]

ValueError: Coefficients arrays must have the same size.
```
