+++ IPython/core/magics/basic.py
@@ -369,6 +369,7 @@ Currently the magic system has the following functions:""",
             shell.InteractiveTB.set_mode(mode=new_mode)
             print('Exception reporting mode:',shell.InteractiveTB.mode)
         except:
+            raise
             xmode_switch_err('user')
 
     @line_magic
